# template-service-parent/template-service-client/target/generated-sources/openapi/api/openapi.yaml
openapi: 3.0.3
info:
  description: API for managing code generation templates
  title: Template Service API
  version: 1.0.0
servers:
- description: Local development server
  url: http://localhost:8080
paths:
  /api/templates:
    get:
      operationId: getAllTemplates
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Template'
                type: array
          description: List of templates
      summary: Get all templates
      x-accepts: application/json
    post:
      operationId: createTemplate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Template'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
          description: Created template
        "400":
          description: Invalid input
      summary: Create a new template
      x-content-type: application/json
      x-accepts: application/json
  /api/templates/{id}:
    delete:
      operationId: deleteTemplate
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Template deleted
      summary: Delete template
      x-accepts: application/json
    get:
      operationId: getTemplateById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
          description: Template found
        "404":
          description: Template not found
      summary: Get template by ID
      x-accepts: application/json
  /api/templates/search:
    get:
      operationId: getTemplate
      parameters:
      - explode: true
        in: query
        name: name
        required: true
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: type
        required: true
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: language
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
          description: Template found
        "404":
          description: Template not found
      summary: "Search template by name, type, and language"
      x-accepts: application/json
  /api/templates/type/{type}/language/{language}:
    get:
      operationId: getTemplatesByTypeAndLanguage
      parameters:
      - explode: false
        in: path
        name: type
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: language
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Template'
                type: array
          description: List of templates
      summary: Get templates by type and language
      x-accepts: application/json
  /api/templates/tag/{tag}:
    get:
      operationId: getTemplatesByTag
      parameters:
      - explode: false
        in: path
        name: tag
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Template'
                type: array
          description: List of templates
      summary: Get templates by tag
      x-accepts: application/json
  /api/templates/seed:
    post:
      operationId: seedTemplates
      responses:
        "200":
          description: Templates seeded successfully
      summary: Seed default templates
      x-accepts: application/json
components:
  schemas:
    Template:
      example:
        name: name
        active: true
        description: description
        language: language
        id: id
        type: type
        version: version
        content: content
        tags:
        - tags
        - tags
      properties:
        id:
          format: objectid
          type: string
        name:
          type: string
        type:
          type: string
        language:
          type: string
        version:
          type: string
        active:
          type: boolean
        description:
          type: string
        tags:
          items:
            type: string
          type: array
        content:
          type: string
      required:
      - content
      - language
      - name
      - type
      type: object